requirements.txt
pandas==2.0.0
openai==1.17.1
azure-search-documents==11.6.0b3
azure-cosmos==4.5.1
seaborn==0.12.2
plotly==5.14.1
streamlit==1.28.1
scikit-learn==1.2.2